+++ src/main/java/com/rdr/sglserverjava/security/TenantContext.java
@@ -2,14 +2,14 @@ package com.rdr.sglserverjava.security;
 
 public final class TenantContext {
 
-    private static final ThreadLocal<Integer> tenantId = new InheritableThreadLocal<>();
+    private static final ThreadLocal<Long> tenantId = new InheritableThreadLocal<>();
 
-    public static void setTenantId(int id) {
+    public static void setTenantId(Long id) {
         System.out.println("setTenantId =" + id);
         tenantId.set(id);
     }
 
-    public static int getTenantId() {
+    public static Long getTenantId() {
         var id = tenantId.get();
         System.out.println("getTenantId =" + id);
         return id;
